# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * LA RC Cafe API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Booking {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  date: string;
  time: string;
  experienceType: string;
  /** @nullable */
  specialRequests?: string | null;
  status: string;
  createdAt: string;
}

export interface BookingInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  email: string;
  phone?: string;
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  time: string;
  /** @minLength 1 */
  experienceType: string;
  specialRequests?: string;
}

export interface BookingUpdate {
  status?: string;
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  date?: string;
  time?: string;
  experienceType?: string;
  specialRequests?: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  priceFrom: number;
  priceUnit: string;
  category: string;
  featured: boolean;
}

export type MenuItemCategory = typeof MenuItemCategory[keyof typeof MenuItemCategory];


export const MenuItemCategory = {
  coffee: 'coffee',
  cold_drinks: 'cold_drinks',
  snacks: 'snacks',
  pizza: 'pizza',
  rc_track: 'rc_track',
  rc_rental: 'rc_rental',
  combo: 'combo',
} as const;

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: MenuItemCategory;
  featured: boolean;
  sortOrder: number;
}

export type MenuItemInputCategory = typeof MenuItemInputCategory[keyof typeof MenuItemInputCategory];


export const MenuItemInputCategory = {
  coffee: 'coffee',
  cold_drinks: 'cold_drinks',
  snacks: 'snacks',
  pizza: 'pizza',
  rc_track: 'rc_track',
  rc_rental: 'rc_rental',
  combo: 'combo',
} as const;

export interface MenuItemInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price: number;
  category: MenuItemInputCategory;
  featured?: boolean;
  sortOrder?: number;
}

export type MenuItemUpdateCategory = typeof MenuItemUpdateCategory[keyof typeof MenuItemUpdateCategory];


export const MenuItemUpdateCategory = {
  coffee: 'coffee',
  cold_drinks: 'cold_drinks',
  snacks: 'snacks',
  pizza: 'pizza',
  rc_track: 'rc_track',
  rc_rental: 'rc_rental',
  combo: 'combo',
} as const;

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: MenuItemUpdateCategory;
  featured?: boolean;
  sortOrder?: number;
}

export interface GalleryItem {
  id: number;
  title: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  subject?: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  message: string;
  createdAt: string;
}

export interface VenueStats {
  totalBookings: number;
  confirmedBookings: number;
  pendingBookings: number;
  totalTracks: number;
  memberCount: number;
  totalProducts: number;
  yearsOpen: number;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  featured: boolean;
  inStock: boolean;
  stock: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price: number;
  /** @minLength 1 */
  category: string;
  featured?: boolean;
  inStock?: boolean;
  stock?: number;
  imageUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  featured?: boolean;
  inStock?: boolean;
  stock?: number;
  imageUrl?: string;
}

export interface AdminLoginInput {
  userId: string;
  password: string;
}

export interface AdminLoginResult {
  token: string;
}

export type VerifyAdmin200 = {
  valid: boolean;
};