# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LA RC Cafe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Bad request
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Cancel/delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, firstName, lastName, email, date, time, experienceType, status, createdAt]
      properties:
        id:               { type: integer }
        firstName:        { type: string }
        lastName:         { type: string }
        email:            { type: string }
        phone:            { type: ["string", "null"] }
        date:             { type: string }
        time:             { type: string }
        experienceType:   { type: string }
        specialRequests:  { type: ["string", "null"] }
        status:           { type: string }
        createdAt:        { type: string }
    BookingInput:
      type: object
      required: [firstName, lastName, email, date, time, experienceType]
      properties:
        firstName:       { type: string, minLength: 1 }
        lastName:        { type: string, minLength: 1 }
        email:           { type: string, minLength: 1 }
        phone:           { type: string }
        date:            { type: string, minLength: 1 }
        time:            { type: string, minLength: 1 }
        experienceType:  { type: string, minLength: 1 }
        specialRequests: { type: string }
    BookingUpdate:
      type: object
      properties:
        status:          { type: string }
        firstName:       { type: string }
        lastName:        { type: string }
        email:           { type: string }
        phone:           { type: string }
        date:            { type: string }
        time:            { type: string }
        experienceType:  { type: string }
        specialRequests: { type: string }
...
    MenuItem:
      type: object
      required: [id, name, description, price, category, featured, sortOrder]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: string }
        price:       { type: integer }
        category:    { type: string, enum: [coffee, cold_drinks, snacks, pizza, rc_track, rc_rental, combo] }
        featured:    { type: boolean }
        sortOrder:   { type: integer }
    MenuItemInput:
      type: object
      required: [name, description, price, category]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
        price:       { type: integer }
        category:    { type: string, enum: [coffee, cold_drinks, snacks, pizza, rc_track, rc_rental, combo] }
        featured:    { type: boolean }
        sortOrder:   { type: integer }
    MenuItemUpdate:
      type: object
      properties:
        name:        { type: string }
        description: { type: string }
        price:       { type: integer }
        category:    { type: string, enum: [coffee, cold_drinks, snacks, pizza, rc_track, rc_rental, combo] }
        featured:    { type: boolean }
        sortOrder:   { type: integer }
...
    AdminLoginInput:
      type: object
      required: [userId, password]
      properties:
        userId:   { type: string }
        password: { type: string }
    AdminLoginResult:
      type: object
      required: [token]
      properties:
        token: { type: string }