# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * LA RC Cafe API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingUpdate {
  status?: string;
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  date?: string;
  time?: string;
  experienceType?: string;
  specialRequests?: string;
}